changes balance from double to int

+++ src/main/java/Transaction.java
@@ -4,11 +4,11 @@ public class Transaction
 {
     private int bank_id; //necessary to facilitate transactions, so we'll have to pass it in somehow (this is the same
                          //as the user's "bank_id"
-    private double amount;
+    private int amount;
     private Date date; //this is automatically set by the database when an entry is made, so this might not be necessary
     
     
-    public Transaction(int bank_id, double amount, Date date)
+    public Transaction(int bank_id, int amount, Date date)
     {
         this.bank_id = bank_id;
         this.amount = amount;
@@ -19,12 +19,12 @@ public class Transaction
         return bank_id;
     }
 
-    public double getAmount(){return amount;}
+    public int getAmount(){return amount;}
     
     public Date getDate(){return date; }
 
    
-    public void setAmount(double amount){
+    public void setAmount(int amount){
         this.amount = amount;
     }
